pyproject.toml
[project]
name = "ai-jounralist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiolimiter>=1.2.1",
    "beautifulsoup4>=4.13.4",
    "elevenlabs>=2.8.1",
    "fastapi>=0.116.1",
    "google-genai>=1.27.0",
    "google-generativeai>=0.8.5",
    "gtts>=2.5.4",
    "langchain-anthropic>=0.3.18",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-google-genai>=2.0.10",
    "langchain-mcp-adapters>=0.1.9",
    "langgraph>=0.6.1",
    "lxml>=6.0.0",
    "mcp>=1.12.2",
    "ollama>=0.5.1",
    "playwright>=1.54.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.47.1",
    "uvicorn>=0.35.0",
]
